no need to assign these, they are unused, spotted by clang

+++ smtpd/smtp.c
@@ -190,7 +190,7 @@ smtp_setup_events(void)
 	purge_config(PURGE_PKI_KEYS);
 
 	maxsessions = ((getdtablesize() - getdtablecount()) & ~0x1) / 2 - SMTP_FD_RESERVE;
-	log_debug("debug: smtp: will accept at most %d clients", maxsessions);
+	log_debug("debug: smtp: will accept at most %d clients", (int)maxsessions);
 
 }
